e2e/util.go: Move waitUntilSize functions to e2eutil for common use
